backend/requirements.txt
django>=2.2.3
django-cors-headers==3.0.1
django-filter==2.1.0
djangorestframework==3.9.4
pylint==1.8.4
pylint-django==2.0.9
pytest==3.5.1